+++ CHANGELOG.md
@@ -9,8 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- README documentation on configurable properties (CIS-3349)
-- README documentation on SMTP and Twilio delivery strategies (AUTHLIB-162)
+- README documentation on configurable properties. (CIS-3349)
+- README documentation on SMTP and Twilio delivery strategies. (AUTHLIB-162)
+- Add configuration property for default email sender address. (CIS-3351)
+- Add configuration property for email subject prefix. (CIS-3351)
+- Prepend email subject prefix to the subject of outgoing email messages. (CIS-3351)
+- **Breaking**: Add a method to the `EmailDeliveryStrategy` interface that sends email using the default sender address. (CIS-3351)
 
 ### Fixed